+++ CHANGELOG.md
@@ -8,6 +8,10 @@
   extract a `use` expression.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- Fixed a bug where the "Extract function" code action would generate a function
+  with the wrong type when used on a use expression.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
 ## v1.13.0-rc1 - 2025-09-29
 
 ### Compiler